move tests into submodule
